requirements.txt
alabaster==0.7.6
astroid==1.3.6
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
## !! Could not determine repository location
bootstrapper==0.1
certifi==2015.11.20.1
colorama==0.3.3
docutils==0.12
ecdsa==0.13
esky==0.9.8
flake8==2.4.1
futures==3.0.3
Jinja2==2.8
libvirt-python==1.3.0
logilab-common==1.0.2
MarkupSafe==0.23
mccabe==0.3.1
msgpack-python==0.4.6
packaging==15.3
paramiko==1.16.0
pep8==1.6.2
py-salt==1.0
pycrypto==2.6.1
pyflakes==0.9.2
Pygments==2.0.2
pylint==1.4.4
pytz==2015.4
pyxdg==0.25
PyYAML==3.11
pyzmq==15.1.0
requests==2.9.1
salt==2015.8.3
singledispatch==3.4.0.3
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
tornado==4.3
Unidecode==0.4.17
virtualenv==13.1.2
wheel==0.24.0
